requirements.txt
# Data Processing
pandas

# Utilities
python-dotenv
pyyaml

# Yahoo Fantasy API
yfpy

# NBA API
nba_api

# LLM and AI
openai
langchain>=1.0.0
langchain-openai
langchain_community
langchain-text-splitters
langgraph

# Youtube Transcript
youtube_transcript_api
google-api-python-client

# Web Scraping
scrapy
requests
firecrawl-py

# Vector Database - PostgreSQL with pgvector (Google Cloud SQL)
langchain-postgres
psycopg2-binary
cloud-sql-python-connector[pg8000,psycopg2]

# Streamlit Web App
streamlit

# Testing
pytest
pytest-cov
pytest-mock
